pyproject.toml
[tool.poetry]
name = "label-gen"
version = "0.1.0"
description = ""
authors = ["Isaias Valle <guitartsword@gmail.com>"]
readme = "README.md"
packages = [{include = "label_gen"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
